==> gwas_code/9_draw_h2.ipynb <==
# write all log file and get the heritability and se
import os
import re
import pandas as pd

eur_h2_logs = "../../../PRS_benchmark/data/result/h2/UKB_h2/"
eas_h2_logs = "../../../PRS_benchmark/data/result/h2/CAS_h2/"
eas_2_h2_log = "../../../PRS_benchmark/data/result/h2/UKB_2_h2/"

def extract_h2_from_logs(folder_path):
    pattern = re.compile(r"Total Observed scale h2:\s*([\d.]+)\s*\(([\d.]+)\)")
    results = []

    if not os.path.exists(folder_path):
        print(f"Folder {folder_path} does not exist.")
        return results

    for filename in os.listdir(folder_path):
        if filename.endswith(".log"):
            file_path = os.path.join(folder_path, filename)
            try:
                with open(file_path, 'r', encoding='utf-8') as f:
                    for line_num, line in enumerate(f, start=1):
                        match = pattern.search(line)
                        if match:
                            h2_value = float(match.group(1))
                            se_value = float(match.group(2))
                            results.append((filename, h2_value, se_value))
            except Exception as e:
                print(f"Error reading {file_path} on line {line_num}: {e}")
    return results

eur_h2_data = extract_h2_from_logs(eur_h2_logs)
eas_h2_data = extract_h2_from_logs(eas_h2_logs)
eas_2_h2_data = extract_h2_from_logs(eas_2_h2_log)

if eur_h2_data:
    print("EUR Heritability Results:")
    for filename, h2, se in eur_h2_data:
        print(f"{filename}: h2 = {h2}, SE = {se}")
    eur_h2_data = pd.DataFrame(eur_h2_data, columns=['filename', 'h2', 'se'])
    eur_h2_data.to_csv("../../../PRS_benchmark/data/result/h2/eur_h2.tsv", sep='\t', index=False)
else:
    print("No EUR heritability results found.")

if eas_h2_data:
    print("EAS Heritability Results:")
    for filename, h2, se in eas_h2_data:
        print(f"{filename}: h2 = {h2}, SE = {se}")
    eas_h2_df = pd.DataFrame(eas_h2_data, columns=['filename', 'h2', 'se'])
    eas_h2_df.to_csv("../../../PRS_benchmark/data/result/h2/eas_h2.tsv", sep='\t', index=False)
else:
    print("No EAS heritability results found.")

if eas_2_h2_data:
    print("EAS to EUR Heritability Results:")
    for filename, h2, se in eas_2_h2_data:
        print(f"{filename}: h2 = {h2}, SE = {se}")
    eas_2_h2_df = pd.DataFrame(eas_2_h2_data, columns=['filename', 'h2', 'se'])
    eas_2_h2_df.to_csv("../../../PRS_benchmark/data/result/h2/eas_2_h2.tsv", sep='\t', index=False)
else:
    print("No EAS heritability results found.")
